+++ src/vesta/__init__.py
@@ -24,4 +24,4 @@ __all__ = (
     "pprint",
 )
 
-__version__ = "1.0.0-dev"
+__version__ = "0.13.0"